config: Move key binding code to a seperate file
